mir-lower: extract shared test scaffolding into test_util module
